tests/requirements.txt
# 测试框架核心
pytest==7.4.3
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.0.0
pytest-timeout==2.2.0
pytest-rerunfailures==13.0
pytest-xdist==3.5.0

# Allure报告
allure-pytest==2.13.2
allure-python-commons==2.13.2

# HTTP请求
requests==2.31.0
urllib3==2.1.0

# 数据处理
PyYAML==6.0.1
jsonschema==4.21.1
faker==22.0.0

# 数据库测试
pymysql==1.1.0
SQLAlchemy==2.0.25

# 性能测试
locust==2.20.0

# 代码质量
pylint==3.0.3
black==24.1.1
isort==5.13.2

# 工具库
python-dotenv==1.0.0
colorama==0.4.6
loguru==0.7.2
